requirements.txt
numpy
tkinter
keras
Pillow
matplotlib